+++ tests/testthat/test-egg_model.R
@@ -14,41 +14,31 @@ test_that("egg_model", {
     )
   )
 
-  expect_no_condition(
-    egg_correlations(
-      fit = res2,
-      period = c(0, 0.5, 1.5, 3.5, 6.5, 10, 12, 17),
-      knots = c(1, 8, 12)
-    )
-  )
-
-  expect_no_condition(
-    egg_aucs(
-      fit = res2,
-      period = c(0, 0.5, 1.5, 3.5, 6.5, 10, 12, 17),
-      knots = c(1, 8, 12)
-    )
-  )
-
-  expect_no_condition(
-    egg_slopes(
-      fit = res2,
-      period = c(0, 0.5, 1.5, 3.5, 6.5, 10, 12, 17),
-      knots = c(1, 8, 12)
-    )
-  )
+  expect_snapshot(egg_correlations(fit = res2)[, lapply(.SD, round, digits = 2L), .SDcols = is.numeric])
+  expect_snapshot({
+    out <- egg_aucs(fit = res2)
+    out[, names(out)[-1L]] <- round(out[, names(out)[-1L]], digits = 2L)
+    out
+  })
+  expect_snapshot({
+    out <- egg_slopes(fit = res2)
+    out[, names(out)[-1L]] <- round(out[, names(out)[-1L]], digits = 2L)
+    out
+  })
 
   for (s in c("predicted", "observed")) {
-    expect_no_condition(compute_apar(fit = res2, from = s)[AP | AR])
+    expect_snapshot({
+      out <- compute_apar(fit = res2, from = s)[AP | AR]
+      out[, c(2L, 3L)] <- round(out[, c(2L, 3L)], digits = 2L)
+      out
+    })
   }
 
-  expect_no_condition(
-    egg_outliers(
-      fit = res2,
-      period = c(0, 0.5, 1.5, 3.5, 6.5, 10, 12, 17),
-      knots = c(1, 8, 12)
-    )
-  )
+  expect_snapshot({
+    out <- egg_outliers(fit = res2)
+    out[, "Distance_IQR"] <- round(out[, "Distance_IQR"], digits = 2L)
+    out
+  })
 
   set.seed(1234)
   dta <- bmigrowth[bmigrowth[["sex"]] == 0, ]
@@ -62,13 +52,16 @@ test_that("egg_model", {
 
   expect_warning(predict_bmi(res)[order(egg_id, egg_ageyears)])
 
-  expect_no_condition(
-    predict_bmi(
+  expect_snapshot({
+    out <- predict_bmi(
       fit = res,
       filter = "source == 'A'"
     )[
       order(egg_id, egg_ageyears),
       egg_bmi := round(egg_bmi, digits = 2)
     ]
-  )
+    out <- head(out, 50)
+    out[, c(3L, 4L)] <- round(out[, c(3L, 4L)], digits = 2L)
+    out
+  })
 })