+++ NEWS.md
@@ -1,3 +1,21 @@
+# eggla 1.0.2
+
+## Fixes
+
+- In `R/*_aucs.R`, `R/*_slopes.R`,
+  - fix: use ID as character and use merge to combine data.frames/matrices.
+
+## Chores
+
+- In `R`, `tests/testthat`,
+  - chore: replace `1:` with `seq_len` and `$` with `[[`.
+
+## Tests
+
+- In `tests/testthat`,
+  - test: cleanup duplicated and outdated tests.
+  - test: add snapshot tests.
+
 # eggla 1.0.1
 
 ## Build